Update gamescan.py
Fixed a bug in gamescan, where only the positive returns under the "numeric positive" class were being shown. Also, in this bug, the daily returns (in percentages) and total returns (in $) were being grouped together. I also added some pandas dataframes, and changed these returns and percentages to be floats, so that we can do some calculations with them later.