Update gamescan.py, Create userscan.py
Added some more things to gamescan. Made it so that it only gathers 1/3 of the users in a game (rounded of course). This is done because usually only the top third of users are even positive. As for userscan, this is the beginning of the program that shall be called when the first round of user data has been gathered (done to save time). It shall analyze the performance tab to see how well they have done for the past 10 days, and shall begin constructing the user ranking based off of this information.