Update gamescan.py, Create playtime.py
Added the ability to find how many days a person has played-- done to find the average of a certain data input over a given range of days (for use in the ranking algorithms). For example, I can find the average amount of trades per day to find the most active players.

I then added this function to gamescan, so that it can be viewed in a pandas dataframe.

Enjoy! :)