Update gamescan.py, Update games.py
Incorporated James's code into gamescan, and took out the nice print statements in games.py as they are no longer needed. The day has finally come where I do not have to wait for the reddit game to be parsed in order to get to actually relevant data. it seems to be that 42 out of the top 145 games are expired... they should really fix that...