Update gamescan.py, Create playtime.py
Had to re-add playtime after the foolish mortal known as James decided to delete it.
As for gamescan, I made it so that the duplicates in games[] are deleted... I would blame it on games.py for having a bug, but it is actually MarketWatch with the bug. For some reason games will show up on multiple pages.