Update gamescan.py
Forgot that we are only taking the top 1/3. When I tested it out on my other machine, it tried to match 7614 items to 3040... Which is strange because that does not equate to 1/3... We shall see how it turns out.

I'll save my customary closing for when another winrar is achieved.