Update gamescan.py
Added a new column- "eff_score". This is a calculated score based off of a user's performance in the past 10 days. It is calculated based off of the differences between those ten days, returning us a total of 9 values.

On some days, there are no changes. On other days, there is either profit or loss. This score is calulcated based off of a user's tendency to do well/not do so well. The formula is ((days-1) + positive_days - negative_days) / 2*(days-1) + zeroes. I shall explain it in more detail in my upcoming documentation-session, when I shall go through and add a bunch of comments.