Update gamescan.py
Added more calculated columns and filters. The ranking system now returns the top 50 based off of an aggregate score called 'mmScore'. This score is calculated based off of a user's ranking relative to the other people in the top 250. If they are first place in all four of the categories, they shall receive a score of 1000, and be the perfect player.

There are two types of categories: scores based off returns, and scores based off player activity. The four categories combined help find the best players of the games we scanned.

TODO: contemplate if we want to expand our crawl to find the best players of the top 300 games instead of the top 150.

as always, enjoy! :)