Create receipt.py
Hello. Quite a bit more added. After waiting about a month for James to get his act together and actually start working on the transactions stuff, I figured that it was time for me to just man up and do it. In his defense, I didn't really know how I was going to go about doing it until I cared enough to try.

Now, I suppose I still am not quite sure how I want to handle these multiple transaction things, but oh well. We shall see where the wind carries us mere withering leaves.

What is currently happening is that when a user makes a transaction, the order factors are combined into a single request ticket, and then hashed through md5 to create somewhat of a 'receipt'. This receipt is then stored in a file known as transactions.log, where we keep track of which orders we have and haven't made so far. Each user will have to keep their own backups of this file, or I'll have to change it to save somewhere else later, as currently it only supports one user at a time.

Most of the more difficult code in this took the time it did because I had to read the documentation for datetime, which was fun. Anyways, as always:

Enjoy! :)