Update mmutils.py, Update buy.py, Update receipt.py, Create jrtools.py
Added lots. Working on the transactions is coming along fairly nicely, though a lot of logic is still to be done for which transactions to process. I have no way of telling orders apart from their counterparts other than thorugh the md5 hashes I am generating as receipt tickets. Also, it seems to be that all order amounts hinge on the comparison of our networths- things that are changing constantly, and thus might not match up when I go to sell holdings that I have purchased earlier.

Overall, I'd say all of this is coming along quite nicely. There is still much to be done, though.